+++ package.json
@@ -10,7 +10,9 @@
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.15.0",
     "react-scripts": "5.0.1",
+    "react-select": "^5.7.4",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
